requirements.txt
django==3.1.1
djangorestframework==3.11.0
mysqlclient==2.0.0
channels==2.4.0
channels_redis==3.1.0
asgiref~=3.2.10